fix(expect-expect): support MemberExpressions in assertFunctionNames
Fixes #175
