+++ rules/expect-expect.js
@@ -6,6 +6,7 @@
  */
 
 const getDocsUrl = require('./util').getDocsUrl;
+const getNodeName = require('./util').getNodeName;
 
 module.exports = {
   meta: {
@@ -27,22 +28,25 @@ module.exports = {
   },
   create(context) {
     const unchecked = [];
-    const assertFunctionNames =
+    const assertFunctionNames = new Set(
       context.options[0] && context.options[0].assertFunctionNames
         ? context.options[0].assertFunctionNames
-        : ['expect'];
+        : ['expect']
+    );
 
     return {
-      'CallExpression[callee.name=/^it|test$/]'(node) {
-        unchecked.push(node);
-      },
-      [`CallExpression[callee.name=/^${assertFunctionNames.join('|')}$/]`]() {
-        // Return early in case of nested `it` statements.
-        for (const ancestor of context.getAncestors()) {
-          const index = unchecked.indexOf(ancestor);
-          if (index !== -1) {
-            unchecked.splice(index, 1);
-            break;
+      CallExpression(node) {
+        const name = getNodeName(node.callee);
+        if (name === 'it' || name === 'test') {
+          unchecked.push(node);
+        } else if (assertFunctionNames.has(name)) {
+          // Return early in case of nested `it` statements.
+          for (const ancestor of context.getAncestors()) {
+            const index = unchecked.indexOf(ancestor);
+            if (index !== -1) {
+              unchecked.splice(index, 1);
+              break;
+            }
           }
         }
       },
